+++ client/src/components/index.ts
@@ -1,1 +1,2 @@
 export { default as Header } from "./shared/header/Header"
+export {default as TextInput} from "./auth/TextInput"
