+++ client/src/pages/Login.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Login = () => {
+  return (
+    <div>Login</div>
+  )
+}
+
+export default Login
